Add first working code.
Voltage on PB3 is measured periodically. LED is turned on if voltage exceed treshold. This will serve as indicator of charged battery. Every cycle PWM is decreaded if voltage is lower than treshold and increaed if voltage id above treshold. This will serve for lowering current when target voltage is achieved.